pulling

+++ EventFiles/interface/DelphesEventFile.hh
@@ -25,7 +25,7 @@ class DelphesEventFile : public EventFile
 
         //virtual GenEventInfoProduct getGenInfo() const override;
         virtual double getMET() const override;
-        virtual long getEventIDNum() const override;
+        virtual unsigned long long getEventIDNum() const override;
         virtual int getNumPileUpInteractions() const override {throw std::runtime_error("not implemented yet");}
         virtual std::vector<bool> getTriggerResults(std::string subProcess) const override {throw std::runtime_error("not implemented yet");}
         virtual std::vector<std::string> getTriggerNames(std::string subProcess) const override {throw std::runtime_error("not implemented yet");}
