+++ Modules/interface/EventInput.hh
@@ -38,7 +38,7 @@ public:
     virtual std::vector<std::string> getTriggerNames(std::string subProcess) const= 0;
     
     virtual double getMET() const = 0;
-    virtual long getEventIDNum() const = 0; 
+    virtual unsigned long long getEventIDNum() const = 0; 
 
     virtual bool checkTrigger(std::string triggerName, std::string subProcess = "") const = 0;
 private:
